requirements.txt
torch==1.9.0
torchvision==0.10.0
opencv-python-headless==4.5.5.64
pillow==8.3.1
pandas
git
ultralytics